Handle nodes that are missing node data
If the node data is missing, this usually means the instance crapped out
during the initial assimilation into Chef. We've seen that occasionally
with EMR nodes when for some reason, EMR decides to not successfully
start the instances, even though the same identical command worked 15
minutes earlier and works again 15 minutes later.

And I know the command is identical, I just uparrowed and hit return.